First step in the implementation of past due date invoices: created a service that computes the due date(s) and outstanding amount(s) based on the Invoice terms; added initial support for multiple payment terms per invoice (e.g. 'pay after 30 days the 50% of the invoice, pay after 60 days the 40%, pay after 90 the 10%'). Part of issue OFBIZ-1530 Modified the getDayEnd util method to be usable in Minilang.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@604922 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/accounting/src/org/ofbiz/accounting/invoice/InvoiceServices.java
@@ -237,7 +237,7 @@ public class InvoiceServices {
             Long orderTermNetDays = orh.getOrderTermNetDays();
             Timestamp dueDate = null;
             if (orderTermNetDays != null) {
-                dueDate = UtilDateTime.getDayEnd(invoiceDate, orderTermNetDays.intValue());
+                dueDate = UtilDateTime.getDayEnd(invoiceDate, orderTermNetDays);
             }
             
             // create the invoice record
@@ -934,7 +934,7 @@ public class InvoiceServices {
                 createInvoiceContext.put("invoiceDate", now);
                 // if there were days associated with the commission agreement, then set a dueDate for the invoice.
                 if (days != null) {
-                    createInvoiceContext.put("dueDate", UtilDateTime.getDayEnd(now, days.intValue()));
+                    createInvoiceContext.put("dueDate", UtilDateTime.getDayEnd(now, days));
                 }
                 createInvoiceContext.put("invoiceTypeId", invoiceType);
                 // start with INVOICE_IN_PROCESS, in the INVOICE_READY we can't change the invoice (or shouldn't be able to...)
